workflow_orchestration/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
yfinance